% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_Hclust.R
\name{all_Hclust}
\alias{all_Hclust}
\title{Silhouette recommendation for hierarchical clustering algorithms}
\usage{
all_Hclust(No_Clusters, dist)
}
\arguments{
\item{No_Clusters}{\code{integer}\cr
Integer value indicating the maximium number of clusters that are considered for calculating the recommendation}

\item{dist}{\code{distance matrix}\cr
Distance matrix of class \code{dist} that is used as input for hierarchical clustering algorithms}
}
\value{
\code{sil_recommendation}\cr
a vector,i.e. the objectively best configuration of: (1) the algorithm, (2) the number of clusters, (3) the respective Silhouette value
}
\description{
This function takes a distance matrix of class \code{dist} as input and iteratively calculates the Silhouette measure for
any potential number of clusters (specified by No_Clusters) and any potential hierarchical cluster algorithm from the set of:
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" , "centroid". The result is a recommendation, indicating
 the best objective configuration, i.e. a vector: \code{sil_recommendation},
with (1) the algorithm, (2) the number of clusters, (3) the respective Silhouette value
}
\seealso{
Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
