% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Tailored data cleaning function}
\usage{
clean_data(dataset)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
Survey dataset to be cleaned}
}
\value{
\code{data}\cr
cleaned survey dataset
}
\description{
This function is tailored to the survey dataset as provided by the University. The reader is refered to chapter 4.2.
In summary, this the main tasks are: normalizations, type conversions, aggregations, deletions, feature extractions
}
\seealso{
Other Helper: \code{\link{discard_meaningless_features}},
  \code{\link{discard_nonrelevant_features}},
  \code{\link{exclude_empirical_correlating_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{get_numeric}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{ordered_to_numeric}},
  \code{\link{remove_attributes}},
  \code{\link{standardize_data}}, \code{\link{tblFun}}
}
