% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_hierarchical.R
\name{clustering_hierarchical}
\alias{clustering_hierarchical}
\title{Wrapper function for hierarchical clustering algorithms}
\usage{
clustering_hierarchical(distance_matrix, method = "ward.D2", number_clusters)
}
\arguments{
\item{distance_matrix}{\code{distance matrix}\cr
Distance matrix of class \code{dist} that is used as input for hierarchical clustering algorithms}

\item{method}{\code{method = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" , "centroid" )}\cr
Selection of the hierarchical clustering algorithm, default is \code{ward.D2}}

\item{number_clusters}{\code{integer}\cr
Integer value determining the number of clusters used by \code{cutree}, i.e. the clustering result will have as many clusters as specified by \code{number_clusters}}
}
\value{
\code{clustering}\cr
Original clustering, i.e. the standard output from \code{hclust}

\code{clusterCut}\cr
Cluster assignment as the result of applying \code{clusterCut} to the original clustering, i.e. the standard output from \code{hclust}
}
\description{
This function wraps the functions \code{hclust} and \code{cutree} and returns a list containing the original clustering as computed by \code{hclust} and
 the resulting cluster assignments based on the passed \code{number_clusters}
}
\seealso{
Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_Hclust}},
  \code{\link{all_PAM}}, \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
