% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_kproto.R
\name{clustering_kproto}
\alias{clustering_kproto}
\title{Wrapper function for kprototypes algorithm}
\usage{
clustering_kproto(x, centers, lambda)
}
\arguments{
\item{x}{\code{data.frame}\cr
dataset to be used for k-prototypes clustering.}

\item{centers}{\code{integer or data.frame}\cr
Either the number of clusters, a vector specifying indices of initial prototypes, or a data frame of prototypes of the same coloumns as x.}

\item{lambda}{\code{vector}\cr
Parameter > 0 to trade off between Euclidean distance of numeric variables and simple matching coefficient between categorical variables.
Also a vector of variable specific factors is possible where the order must correspond to the order of the variables in the data.
In this case all variables' distances will be multiplied by their corresponding lambda value.}
}
\value{
\code{clusterCut}\cr
Resulting cluster assignment
}
\description{
This function wraps the function \code{kproto.default} and returns a vector of the resulting cluster assignments.
}
\seealso{
The reader is referred to the original documentation of the k prototypes clustering algorithm \url{https://cran.r-project.org/web/packages/clustMixType/clustMixType.pdf}

Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_Hclust}},
  \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
