% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_pam.R
\name{clustering_pam}
\alias{clustering_pam}
\title{Wrapper function for partitioning around medoids (PAM) algorithm}
\usage{
clustering_pam(distance_matrix, number_clusters)
}
\arguments{
\item{distance_matrix}{\code{distance matrix}\cr
Distance matrix of class \code{dist} that is used as input for PAM clustering algorithms}

\item{number_clusters}{\code{vector}\cr
positive integer specifying the number of clusters, less than the number of observations.}
}
\value{
\code{clusterCut}\cr
Resulting cluster assignment
}
\description{
This function wraps the function \code{pam} from the package \link[cluster]{pam} and returns a vector of the resulting cluster assignments.
}
\seealso{
The reader is referred to the original documentation of the pam clustering algorithm \url{https://cran.r-project.org/web/packages/cluster/cluster.pdf}

Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_Hclust}},
  \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
