% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{Correlations}
\alias{Correlations}
\title{Correlation heatmap}
\usage{
Correlations(dataset, threshold = 0, threshold_count = 0, p.value = 0,
  method = "pearson")
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The survey dataframe for which correations shall be computed}

\item{threshold}{\code{data.frame}\cr
minimum required degree of correlation to be shown in the heatmap visualisation}

\item{threshold_count}{\code{data.frame}\cr
required number of correlations with other features above the threshold in order to be shown in the heatmap visualisation}

\item{p.value}{\code{data.frame}\cr
required signicifance level for correlation significance test to be shown in the heatmap visualisation}
}
\description{
This function takes the survey data frame as input and outputs a heatmap visualisation of the correlation matrix (folder: "./Output/Images/correlations/"). The user
passes the survey dataset and can specifies a threshold (minimum degree of correlation), a threshold count (number of correlations with other features above the threshold),
a p.value (signicifance level for correlation significance test) and a method (either pearson or spearman)
}
\seealso{
Other Additional: \code{\link{Qualitative_statistics}},
  \code{\link{Quantitative_statistics}},
  \code{\link{Summaries}}, \code{\link{decision_tree}}
}
