% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree.R
\name{decision_tree}
\alias{decision_tree}
\title{Decision tree for feature selection}
\usage{
decision_tree(tree_data)
}
\arguments{
\item{dataset}{\code{tree_data}\cr
The supplied data frame named "tree_data"}
}
\description{
This function only works for the provided classification data (tree_data.rds). It produces the rpart decision tree
that is displayed in chapter 4.2 in the thesis' document. The decision tree is computed based on the identified set of unweighted, relevant features
(CD04, CD05, CD06, CD02 and CD08C). The reader is referred to chapter 4.2 for further details.
}
\seealso{
Other Additional: \code{\link{Correlations}},
  \code{\link{Qualitative_statistics}},
  \code{\link{Quantitative_statistics}},
  \code{\link{Summaries}}
}
