% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_nonrelevant_features.R
\name{discard_nonrelevant_features}
\alias{discard_nonrelevant_features}
\title{Simple function that excludes non-relevant features}
\usage{
discard_nonrelevant_features(data)
}
\arguments{
\item{data}{\code{data.frame}\cr
The survey dataset from which non-relevant features are to be discarded}
}
\description{
This function simply removes columns that are not considered as relevant. Those columns or features are not loaded into the shiny application.
}
\seealso{
Other Helper: \code{\link{clean_data}},
  \code{\link{discard_meaningless_features}},
  \code{\link{exclude_empirical_correlating_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{get_numeric}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{ordered_to_numeric}},
  \code{\link{remove_attributes}},
  \code{\link{standardize_data}}, \code{\link{tblFun}}
}
