% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_gower_subjects_function.R
\name{dist_gower_subjects_function}
\alias{dist_gower_subjects_function}
\title{Wrapper function for the Gower coefficient}
\usage{
dist_gower_subjects_function(data, weights = NULL, asym.bin = NULL,
  alwaysGower = FALSE)
}
\arguments{
\item{data}{\code{data.frame}\cr
From the original documentation: matrix or data frame containing the variables. Variables can be numeric, ordered, or factor.
Symmetric or asymmetric binary variables should be numeric and only contain 0 and 1.
character variables will be converted to factor. NAs are tolerated.}

\item{weights}{\code{vector}\cr
vector listing the weights for the variables in x. Can be missing, in which case all variables have equal weights.}

\item{asym.bin}{\code{vector}\cr
vector listing the asymmetric binary variables in x}

\item{alwaysGower}{\code{logical}\cr
logical indicating if the Gower coefficient should always be used or not. If TRUE, and only numeric features supplied, then euclidean distances are calculated}
}
\value{
\code{Distance_matrix}\cr
Resulting distance matrix
}
\description{
This function wraps the function \code{gowdis} from the package \link[FD]{gowdis} and returns a distance matrix as a result of the application of the
Gower coefficient. The reader is referred to the details as provided in \link[FD]{gowdis}.
}
\references{
Podani, J. (1999) Extending Gower's general coefficient of similarity to ordinal characters. Taxon 48:331-340.
}
\seealso{
Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_Hclust}},
  \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
