% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_empirical_correlating_features.R
\name{exclude_empirical_correlating_features}
\alias{exclude_empirical_correlating_features}
\title{Simple function that excludes empirically correlating features}
\usage{
exclude_empirical_correlating_features(data)
}
\arguments{
\item{data}{\code{data.frame}\cr
The survey dataset from which non-relevant features are to be discarded}
}
\description{
This function simply removes columns that are empirically correlating to other features.
Empirically correlating features are one sort of inadmissible features as defined in the thesis' document in chapter 2.3.2.3.
The corresponding correlation matrix can be found in appendix A.
Those columns or features are not loaded into the shiny application for the primary data analysis.
Empirically correlating features will only be included in the tab "Summary".
}
\seealso{
Other Helper: \code{\link{clean_data}},
  \code{\link{discard_meaningless_features}},
  \code{\link{discard_nonrelevant_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{get_numeric}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{ordered_to_numeric}},
  \code{\link{remove_attributes}},
  \code{\link{standardize_data}}, \code{\link{tblFun}}
}
