% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric.R
\name{get_numeric}
\alias{get_numeric}
\title{Simple function extracting numeric features and indices}
\usage{
get_numeric(dataset)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The dataframe from which numeric features and indices shall be extracted}
}
\value{
\code{dataset}\cr
A reduced data set with only numeric columns

\code{cols}\cr
A vector with the index of all numeric columns as in the original data frame
}
\description{
This function takes a data frame as input and returns a list with a reduced data set with only numeric columns and a vector with their index in the
original data frame
}
\seealso{
Other Helper: \code{\link{clean_data}},
  \code{\link{discard_meaningless_features}},
  \code{\link{discard_nonrelevant_features}},
  \code{\link{exclude_empirical_correlating_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{ordered_to_numeric}},
  \code{\link{remove_attributes}},
  \code{\link{standardize_data}}, \code{\link{tblFun}}
}
