% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\name{Help}
\alias{Help}
\title{User interface for help page}
\usage{
Help()
}
\description{
This function implements all UI elements under the tab "Help". Implementations are based on the functionality provided by
the shiny R package and shinydashboard R package.
}
\seealso{
Other UI: \code{\link{Start}}, \code{\link{Summary_ui}},
  \code{\link{Taxonomy_ui}}, \code{\link{mainPage}},
  \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabDesc}},
  \code{\link{tabPwT}}
}
