% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainPage.R
\name{mainPage}
\alias{mainPage}
\title{Modified navbarPage from shiny}
\usage{
mainPage(title, ..., id = "dashboard", thead = NULL, header = NULL,
  footer = NULL, windowTitle = title)
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Summary_ui}}, \code{\link{Taxonomy_ui}},
  \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabDesc}},
  \code{\link{tabPwT}}
}
