% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidimensional_scaling.R
\name{multidimensional_scaling}
\alias{multidimensional_scaling}
\title{Multidimensional scaling for cluster visualisation}
\usage{
multidimensional_scaling(distance_matrix, cluster = NULL, k = 2)
}
\arguments{
\item{distance_matrix}{\code{distance matrix}\cr
From the original documentation: a distance structure such as that returned by dist or a full symmetric matrix containing the dissimilarities.
character variables will be converted to factor. NAs are tolerated.}

\item{cluster}{\code{vector}\cr
A clustering result according to which the MDS result is being split}

\item{k}{\code{integer}\cr
the maximum dimension of the space which the data are to be represented in; must be in {1, 2, …, n-1}.}
}
\value{
\code{cluster_groups}\cr
A list containing either one or multiple data frames (depending on the value of \code{cluster}), that represent the initial distances from the \code{distance matrix}
in k-dimensional space
}
\description{
This function applies the technique of multidimensional scaling (MDS) based on a distance matrix. The user may pass a cluster assignment, which leads
to the MDS result being split accordingly. In other words, the result is either one data frame of \code{k} columns (the new dimensions resulting from applying MDS), or multiple
dataframes in a list of data frames depending on the number of clusters from the provided cluster assignment \code{cluster}. For details about the MDS, the reader is referred to
\link[stats]{cmdscale}
}
\seealso{
Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_Hclust}},
  \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{pca_clustering}}
}
