% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_to_numeric.R
\name{ordered_to_numeric}
\alias{ordered_to_numeric}
\title{Simple function converting ordered features to numeric features}
\usage{
ordered_to_numeric(dataset)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The dataframe of which ordered features shall be converted to numeric features}
}
\value{
\code{dataset}\cr
A dataset that has all initally features of class \code{ordered} converted to features of class \code{numeric}
}
\description{
This function takes a data frame as input and returns a data frame that has all initally features of class \code{ordered} converted to features of class \code{numeric}
}
\seealso{
Other Helper: \code{\link{clean_data}},
  \code{\link{discard_meaningless_features}},
  \code{\link{discard_nonrelevant_features}},
  \code{\link{exclude_empirical_correlating_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{get_numeric}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{remove_attributes}},
  \code{\link{standardize_data}}, \code{\link{tblFun}}
}
