% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_clustering.R
\name{pca_clustering}
\alias{pca_clustering}
\title{Principle component analysis for cluster visualisation}
\usage{
pca_clustering(cluster, dataframe, weights = NULL)
}
\arguments{
\item{cluster}{\code{vector}\cr
A clustering result according to which the PCA result is being split}

\item{dataframe}{\code{data.frame}\cr
A data frame consisting quantitative as well as qualitative data to be processes by \link[PCAmixdata]{PCAmix}}

\item{weights}{\code{vector}\cr
A vector of weights for both, the quantitative as well as qualitative features}
}
\value{
\code{cluster_groups}\cr
A list containing either one or multiple data frames (depending on the value of \code{cluster}), that represesent the supplied data frame in two-dimensional space.

\code{dims}\cr
Additionally, a vector containing the eigenvalues of the first two principle components is returned, indicating the explained variance of the first components.
}
\description{
This function applies the technique of principle component analysis (PCA) based on a data frame of mixed data.
The reader is referred to the R function \link[PCAmixdata]{PCAmix}, on which the PCA implementation is based. The user may pass a cluster assignment, which leads
to the PCA result being split accordingly. In other words, the result is either one data frame of \code{k} columns (the new dimensions resulting from applying PCA), or multiple
dataframes in a list of data frames depending on the number of clusters from the provided cluster assignment \code{cluster}. For details about the PCA, the reader is referred to
\link[stats]{cmdscale}
}
\seealso{
Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{Silhouettes}}, \code{\link{all_Hclust}},
  \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}}
}
