% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_feature_selection.R
\name{PCA_feature_selection}
\alias{PCA_feature_selection}
\title{Principle component analysis for feature selection}
\usage{
PCA_feature_selection(dataframe, no_components,
  absolute_relative = c("absolute", "relative"), no_abs_features = 2,
  weights = NULL)
}
\arguments{
\item{dataframe}{\code{data.frame}\cr
A data frame consisting of quantitative and qualitative features}

\item{no_components}{\code{integer}\cr
The maximal number of components to be considered}

\item{absolute_relative}{\code{absolute_relative = c("absolute", "relative")}\cr
a character indicating whether features are extracted based on an absolute or on a relative measurement of feature contributions.}

\item{no_abs_features}{\code{integer}\cr
Integeger value defining the number of principle components from which features can be potentially extracted}

\item{weights}{\code{vector}\cr
A vector of weights for both quantitative and qualitative features}
}
\value{
\code{selected_cols}\cr
A vector containing selected features (absolute order across components)

\code{selected_cols_aggregated}\cr
A vector containing selected features (aggregated over all considered principle component)
}
\description{
This function applies data-driven principle component (PCA) based filtering technique for feature selection. The PCA implementation is based on \link[PCAmixdata]{PCAmix}. The reader is referred to chapter 4.2, for a detailed
justification and reasoning of this approach. Features are extracted from a predefined number of principle components either based on an absolute or on a relative measurement of feature contributions.
 Feature contributions can be aggregated over all considered principle component or put in an absolute order across components.
}
\seealso{
Other Clustering: \code{\link{Silhouettes}},
  \code{\link{all_Hclust}}, \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
