% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qualitative_statistics.R
\name{Qualitative_statistics}
\alias{Qualitative_statistics}
\title{Survey statistics for qualitative features}
\usage{
Qualitative_statistics(dataset)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The survey dataframe for which the statistics shall be computed}
}
\description{
This function takes the survey data frame as input and produces a summary sheet for each qualitative feature.
This sheet includes a table reporting on simple measures: feature, number of nominal levels, number of entries, ratio between number of levels and number of entries (rows), number of NAs
Additionally, a bar-chart shows the underlying distribution of feature values with associated counts and relative proportions
}
\seealso{
Other Additional: \code{\link{Correlations}},
  \code{\link{Quantitative_statistics}},
  \code{\link{Summaries}}, \code{\link{decision_tree}}
}
