% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantitative_statistics.R
\name{Quantitative_statistics}
\alias{Quantitative_statistics}
\title{Survey statistics for quantitative features}
\usage{
Quantitative_statistics(dataset)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The survey dataframe for which the statistics shall be computed}
}
\description{
This function takes the survey data frame as input and produces a summary sheet for each quantitative feature.
This sheet includes a table reporting on simple measures: feature, min, max, mean, median, vairance, standard deviation, first quartile, third quartile, interquartilerange.
Additionally, a histogram shows the underlying distribution of feature values.
}
\seealso{
Other Additional: \code{\link{Correlations}},
  \code{\link{Qualitative_statistics}},
  \code{\link{Summaries}}, \code{\link{decision_tree}}
}
