% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_attributes.R
\name{remove_attributes}
\alias{remove_attributes}
\title{Simple function that removes all attributes associated with a feature}
\usage{
remove_attributes(dataset)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The dataframe of which the attributes shall be removed}
}
\value{
\code{dataset}\cr
A dataset without attributes
}
\description{
This function simply removes any attributes from the features of a data frame (comments, class, ...)
}
\seealso{
Other Helper: \code{\link{clean_data}},
  \code{\link{discard_meaningless_features}},
  \code{\link{discard_nonrelevant_features}},
  \code{\link{exclude_empirical_correlating_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{get_numeric}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{ordered_to_numeric}},
  \code{\link{standardize_data}}, \code{\link{tblFun}}
}
