% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectizeInputF.R
\name{selectizeInputF}
\alias{selectizeInputF}
\title{Wrapper function for selectizeInput}
\usage{
selectizeInputF(inputId, ..., label = NULL, options, width)
}
\description{
A simple wrapper function that allows selectizeInput with selectize = FALSE
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Summary_ui}}, \code{\link{Taxonomy_ui}},
  \code{\link{mainPage}}, \code{\link{silhouette_viz}},
  \code{\link{tabDesc}}, \code{\link{tabPwT}}
}
