% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server_logic.R
\name{Server_logic}
\alias{Server_logic}
\title{Server logic of shiny web application}
\usage{
Server_logic(input, output, session, data)
}
\description{
This function implements the complete server logic. All calculations (MDS-visualization, PCA visualisation,
silhouette measure calculation, silhouette visualization, all table outputs, recommendation calculation, improvement calculation)
are handled by this server function. The server function reacts to user inputs as specified in \link[Taxonomy]{Taxonomy_ui} and \link[Taxonomy]{Summary_ui}.
}
