% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette_viz.R
\name{silhouette_viz}
\alias{silhouette_viz}
\title{Silhouette display with standard color palette}
\usage{
silhouette_viz(sil.obj, label = FALSE, return.summary = TRUE, ...)
}
\description{
This function is based on the function \link[factoextra]{fviz_silhouette}. It visualizes a silhouette object and applies the standard color palette (used for the shiny application)
The reader is referred to the original documentation \link[factoextra]{fviz_silhouette}.
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Summary_ui}}, \code{\link{Taxonomy_ui}},
  \code{\link{mainPage}}, \code{\link{selectizeInputF}},
  \code{\link{tabDesc}}, \code{\link{tabPwT}}
}
