% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silhouettes.R
\name{Silhouettes}
\alias{Silhouettes}
\title{Wrapper function for silhouette measure computation}
\usage{
Silhouettes(Clustering, dist)
}
\arguments{
\item{Clustering}{\code{vector}\cr
A cluster assignment as a result from any clustering algorithm}

\item{dist}{\code{distance matrix}\cr
Distance matrix of class \code{dist} that is used as input for silhouette measure calculation}
}
\value{
\code{Silhouette}\cr
data frame with summary statistics including: cluster number, cluster size, intra-cluster silhouette width/measure

\code{avg_sil}\cr
Overall absolute silhouette width/measure
}
\description{
This function wraps the function \link[cluster]{silhouette}. It takes a distance matrix and a clustering solution as input
 and returns a list containing summary statistics, i.e. silhouette width/measure for individual clusters and the across all clusters (average silhouette widht/measure)
 the resulting cluster assignments based on the passed \code{number_clusters}
}
\seealso{
Other Clustering: \code{\link{PCA_feature_selection}},
  \code{\link{all_Hclust}}, \code{\link{all_PAM}},
  \code{\link{clustering_hierarchical}},
  \code{\link{clustering_kproto}},
  \code{\link{clustering_pam}},
  \code{\link{dist_gower_subjects_function}},
  \code{\link{multidimensional_scaling}},
  \code{\link{pca_clustering}}
}
