% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start.R
\name{Start}
\alias{Start}
\title{User interface for front page}
\usage{
Start()
}
\description{
This function implements all UI elements under the tab "Start". Implementations are based on the functionality provided by
the shiny R package and shinydashboard R package.
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Summary_ui}},
  \code{\link{Taxonomy_ui}}, \code{\link{mainPage}},
  \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabDesc}},
  \code{\link{tabPwT}}
}
