% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{Summaries}
\alias{Summaries}
\title{Survey summary}
\usage{
Summaries(dataset, filename)
}
\arguments{
\item{dataset}{\code{data.frame}\cr
The survey dataframe for which the summary shall be computed}
}
\description{
This function takes the survey data frame as input and writes a .csv file to the current working directory reporting on data type, column number,
number of empty cells, number of NAs, feature comments
}
\seealso{
Other Additional: \code{\link{Correlations}},
  \code{\link{Qualitative_statistics}},
  \code{\link{Quantitative_statistics}},
  \code{\link{decision_tree}}
}
