% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_ui.R
\name{Summary_ui}
\alias{Summary_ui}
\title{User interface for summary functionality}
\usage{
Summary_ui()
}
\description{
This function implements all UI elements under the tab "Summary". Implementations are based on the functionality provided by
the shiny R package and shinydashboard R package.
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Taxonomy_ui}}, \code{\link{mainPage}},
  \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabDesc}},
  \code{\link{tabPwT}}
}
