% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabDesc.R
\name{tabDesc}
\alias{tabDesc}
\title{Function handling tab descriptions}
\usage{
tabDesc(x)
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Summary_ui}}, \code{\link{Taxonomy_ui}},
  \code{\link{mainPage}}, \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabPwT}}
}
