% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabPwT.R
\name{tabPwT}
\alias{tabPwT}
\title{Modified tabs}
\usage{
tabPwT(title, ...)
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Summary_ui}}, \code{\link{Taxonomy_ui}},
  \code{\link{mainPage}}, \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabDesc}}
}
