% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxonomy_ui.R
\name{Taxonomy_ui}
\alias{Taxonomy_ui}
\title{User interface for interactive clustering}
\usage{
Taxonomy_ui()
}
\description{
This function implements all UI elements under the tab "Taxonomy". Implementations are based on the functionality provided by
the shiny R package and shinydashboard R package.
}
\seealso{
Other UI: \code{\link{Help}}, \code{\link{Start}},
  \code{\link{Summary_ui}}, \code{\link{mainPage}},
  \code{\link{selectizeInputF}},
  \code{\link{silhouette_viz}}, \code{\link{tabDesc}},
  \code{\link{tabPwT}}
}
