% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblFun.R
\name{tblFun}
\alias{tblFun}
\title{Simple function to compute percentage statisticss}
\usage{
tblFun(x)
}
\arguments{
\item{x}{\code{vector}\cr
Vector with qualitative feature values, such as feature of class \code{factor}\cr}
}
\value{
\code{res}\cr
Table reporting on value counts and corresponding relative proportions

\code{selected_cols_aggregated}\cr
A vector containing selected features (aggregated over all considered principle component)
}
\description{
This function extends the functionality of the built-in \link[base]{table} function as it adds the relative proportions of the features compared to all other features
}
\seealso{
Other Helper: \code{\link{clean_data}},
  \code{\link{discard_meaningless_features}},
  \code{\link{discard_nonrelevant_features}},
  \code{\link{exclude_empirical_correlating_features}},
  \code{\link{exclude_meaningless_nonrelevant_features}},
  \code{\link{get_numeric}},
  \code{\link{logical_to_factor}},
  \code{\link{logical_to_numerical}},
  \code{\link{ordered_to_factor}},
  \code{\link{ordered_to_integer}},
  \code{\link{ordered_to_numeric}},
  \code{\link{remove_attributes}},
  \code{\link{standardize_data}}
}
